requirements.txt
asgiref==3.3.4
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cloudpickle==1.6.0
cryptography==3.4.7
cycler==0.10.0
dataclasses==0.8
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.0.7
django-allauth==0.44.0
django-crispy-forms==1.8.1
django-storages==1.11.1
google-api-core==1.27.0
google-auth==1.30.0
google-cloud-core==1.6.0
google-cloud-storage==1.38.0
google-crc32c==1.1.2
google-resumable-media==1.3.0
googleapis-common-protos==1.53.0
idna==2.10
importlib-metadata==4.0.1
install==1.3.4
kiwisolver==1.3.1
matplotlib==3.0.3
numpy==1.16.6
oauthlib==3.1.0
opencv-python==4.1.0.25
opencv-python-headless==4.5.2.52
packaging==20.9
Pillow==8.2.0
protobuf==3.17.0
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
termcolor==1.1.0
torch==1.4.0
torchvision==0.5.0
typing-extensions==3.10.0.0
urllib3==1.26.4
whitenoise==5.2.0
yacs==0.1.8
zipp==3.4.1
